refactor(models): split performance comparison entities
- Keep each SeaORM entity's required prelude imports within its own module
- Preserve existing module paths, fields, and entity relationships
